Add getImageByID to each card component.
